pyproject.toml
[project]
name = "role-mentor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asqlite>=2.0.0",
    "browser-use>=0.7.10",
    "fastapi>=0.118.0",
    "groq>=0.31.1",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.3.9",
    "langfuse>=3.6.1",
    "langgraph>=0.6.8",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "lmstudio>=1.5.0",
    "mem0ai>=0.1.118",
    "mermaid-python>=0.1",
    "modal>=1.1.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "streamlit-pdf>=1.0.7",
    "uvicorn>=0.37.0",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[project.scripts]
frontend = "src.frontend.app:main"
backend = "src.backend.api.app:main"
